Use Grid component instead of Table component
